Move AuthorityService to a separate module
